requirements.txt
atproto>=0.0.54
APScheduler>=3.10.4
click>=8.1.7
python-dotenv>=1.0.0
rich>=13.7.0
